requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
python-docx>=1.0.1
langdetect>=1.0.9 